refactor: remove postfix(components)
