+++ app/contact/page.tsx
@@ -1,7 +1,7 @@
 "use client";
 
-import Loading from "@/src/components/common/Loading.component";
-import Contact from "@/src/components/contact/Contact.component";
+import Loading from "@/src/components/common/Loading";
+import Contact from "@/src/components/contact/Contact";
 import { loadingState } from "@/src/stores/common";
 import { useAtom } from "jotai";
 import { useEffect } from "react";
