+++ src/components/main/Award/GroupItem.tsx
