+++ src/components/main/RecruitFloat.tsx
@@ -2,7 +2,7 @@ import { RECRUIT, RECRUIT_FLOAT } from "@/src/constants/recruit/recruit.ko";
 import gsap from "gsap";
 import Image from "next/image";
 import { useRef } from "react";
-import LinkTo from "../common/LinkTo.component";
+import LinkTo from "../common/LinkTo";
 import useTimeDiffer from "@/src/hooks/useTimeDiffer";
 
 const RecruitFloat = () => {
