+++ src/components/portfolio/PortfolioDetailDescription.tsx
