+++ src/components/recruit/Recruit.tsx
@@ -1,9 +1,9 @@
 import { RECRUIT } from "@/src/constants/recruit/recruit.ko";
-import LinkTo from "../common/LinkTo.component";
+import LinkTo from "../common/LinkTo";
 import { ScrollTrigger } from "gsap/ScrollTrigger";
 import { useEffect } from "react";
 import gsap from "gsap";
-import HambergerMenu from "../common/Hamberger.component";
+import HambergerMenu from "../common/Hamberger";
 
 gsap.registerPlugin(ScrollTrigger);
 const Recruit = () => {
